api/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
appdirs==1.4.4            # via black
appnope==0.1.0            # via ipython
asgiref==3.2.10           # via django
attrs==19.3.0             # via black, pytest
backcall==0.2.0           # via ipython
bandit==1.6.2             # via -r requirements/dev.in, pytest-bandit
black==19.10b0            # via -r requirements/dev.in
click==7.1.2              # via black, pip-tools
coverage==5.1             # via pytest-cov
decorator==4.4.2          # via ipython, traitlets
django-debug-toolbar==2.2  # via -r requirements/dev.in
django-environ==0.4.5     # via -r requirements/common.in
django-extensions==2.2.9  # via -r requirements/dev.in
django-filter==2.3.0      # via -r requirements/common.in
django==3.0.7             # via -r requirements/common.in, django-debug-toolbar, django-filter, djangorestframework
djangorestframework==3.11.0  # via -r requirements/common.in
gitdb==4.0.5              # via gitpython
gitpython==3.1.3          # via bandit
gunicorn==20.0.4          # via -r requirements/common.in
ipdb==0.13.2              # via -r requirements/dev.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.15.0           # via -r requirements/dev.in, ipdb
jedi==0.17.1              # via ipython
mccabe==0.6.1             # via pylama
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest
parso==0.7.0              # via jedi
pathspec==0.8.0           # via black
pbr==5.4.5                # via stevedore
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.2.1          # via -r requirements/dev.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.5     # via ipython
psycopg2-binary==2.8.5    # via -r requirements/common.in
ptyprocess==0.6.0         # via pexpect
py==1.8.2                 # via pytest
pycodestyle==2.6.0        # via pylama
pydocstyle==5.0.2         # via pylama
pyflakes==2.2.0           # via pylama
pygments==2.6.1           # via ipython
pylama==7.7.1             # via -r requirements/dev.in
pyparsing==2.4.7          # via packaging
pytest-bandit==0.5.2      # via -r requirements/dev.in
pytest-cov==2.10.0        # via -r requirements/dev.in
pytest-django==3.9.0      # via -r requirements/dev.in
pytest==5.4.3             # via -r requirements/dev.in, pytest-bandit, pytest-cov, pytest-django
pytz==2020.1              # via django
pyyaml==5.3.1             # via bandit
regex==2020.6.8           # via black
six==1.15.0               # via bandit, django-extensions, packaging, pip-tools, traitlets
smmap==3.0.4              # via gitdb
snowballstemmer==2.0.0    # via pydocstyle
sqlparse==0.3.1           # via django, django-debug-toolbar
stevedore==2.0.1          # via bandit
toml==0.10.1              # via black
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black
wcwidth==0.2.4            # via prompt-toolkit, pytest
werkzeug==1.0.1           # via -r requirements/dev.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
